# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LifeLoom API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organ-types": {
            "get": {
                "tags": [
                    "Tipos de Órgãos"
                ],
                "summary": "Listar todos os tipos de órgãos",
                "operationId": "0121716b7375e6d4a85de8defa56bbaf",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de órgãos retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipos de órgãos listados com sucesso."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "id": 1,
                                                    "name": "Coração",
                                                    "description": "Órgão vital",
                                                    "default_expiration_days": 7,
                                                    "default_distance_limit": 500,
                                                    "compatibility_criteria": "",
                                                    "is_post_mortem": true
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de Órgãos"
                ],
                "summary": "Criar um novo tipo de órgão",
                "operationId": "4625255b68dbb31632ab47ecf621a31f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "default_expiration_days",
                                    "default_distance_limit"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Coração"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Órgão vital"
                                    },
                                    "default_expiration_days": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "default_distance_limit": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "compatibility_criteria": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "is_post_mortem": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de órgão criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de órgão criado com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Coração",
                                                "description": "Órgão vital",
                                                "default_expiration_days": 7,
                                                "default_distance_limit": 500,
                                                "compatibility_criteria": "",
                                                "is_post_mortem": true
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organ-types/{id}": {
            "get": {
                "tags": [
                    "Tipos de Órgãos"
                ],
                "summary": "Obter detalhes de um tipo de órgão",
                "operationId": "3d99f9873b4848400f497d994f4dcd63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de órgão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de órgão retornado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de órgão obtido com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Coração",
                                                "description": "Órgão vital",
                                                "default_expiration_days": 7,
                                                "default_distance_limit": 500,
                                                "compatibility_criteria": "",
                                                "is_post_mortem": true
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de órgão não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de órgão não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Órgãos"
                ],
                "summary": "Atualizar um tipo de órgão existente",
                "operationId": "3ccee6e352abf3bba12a5e6909794746",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de órgão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Coração Atualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Órgão vital atualizado"
                                    },
                                    "default_expiration_days": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "default_distance_limit": {
                                        "type": "integer",
                                        "example": 600
                                    },
                                    "compatibility_criteria": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "is_post_mortem": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de órgão atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tipo de órgão atualizado com sucesso."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Coração Atualizado",
                                                "description": "Órgão vital atualizado",
                                                "default_expiration_days": 8,
                                                "default_distance_limit": 600,
                                                "compatibility_criteria": "",
                                                "is_post_mortem": false
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}